chore(plugins): remove dead hook runner and dedupe plugin loader
The plugins crate carried a second HookRunner implementation
(plugins/src/hooks.rs) that had zero consumers — the runtime crate
defines its own HookRunner/HookRunResult/HookEvent with richer
exit-code/stdout-JSON semantics, which is what actually executes plugin
hooks. Delete the dead file and its re-exports so hook behavior has a
single source of truth.

Also collapse the two byte-identical plugin-loader assemblies
(main.rs build_plugin_manager and tools.rs load_plugin_outcome_for_cwd,
each with its own resolve_plugin_path) into a single
RuntimePluginConfig::to_plugin_manager_config method in the runtime
crate.
