+++ rust/crates/plugins/src/lib.rs
@@ -1,4 +1,3 @@
-mod hooks;
 #[cfg(test)]
 pub mod test_isolation;
 
@@ -12,8 +11,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
-
-pub use hooks::{HookEvent, HookRunResult, HookRunner};
 
 const EXTERNAL_MARKETPLACE: &str = "external";
 const BUILTIN_MARKETPLACE: &str = "builtin";
