+++ rust/crates/rusty-sudocode-cli/src/main.rs
@@ -99,8 +99,7 @@ use compat_harness::{extract_manifest, UpstreamPaths};
 use dialoguer::{FuzzySelect, Select};
 use init::initialize_repo;
 use plugins::{
-    render_plugin_capabilities_section, PluginLoadOutcome, PluginManager, PluginManagerConfig,
-    PluginRegistry,
+    render_plugin_capabilities_section, PluginLoadOutcome, PluginManager, PluginRegistry,
 };
 use render::{MarkdownStreamState, Spinner, TerminalRenderer};
 use runtime::{
@@ -3913,35 +3912,10 @@ fn build_plugin_manager(
     loader: &ConfigLoader,
     runtime_config: &runtime::RuntimeConfig,
 ) -> PluginManager {
-    let plugin_settings = runtime_config.plugins();
-    let mut plugin_config = PluginManagerConfig::new(loader.config_home().to_path_buf());
-    plugin_config.enabled_plugins = plugin_settings.enabled_plugins().clone();
-    plugin_config.external_dirs = plugin_settings
-        .external_directories()
-        .iter()
-        .map(|path| resolve_plugin_path(cwd, loader.config_home(), path))
-        .collect();
-    plugin_config.install_root = plugin_settings
-        .install_root()
-        .map(|path| resolve_plugin_path(cwd, loader.config_home(), path));
-    plugin_config.registry_path = plugin_settings
-        .registry_path()
-        .map(|path| resolve_plugin_path(cwd, loader.config_home(), path));
-    plugin_config.bundled_root = plugin_settings
-        .bundled_root()
-        .map(|path| resolve_plugin_path(cwd, loader.config_home(), path));
+    let plugin_config = runtime_config
+        .plugins()
+        .to_plugin_manager_config(cwd, loader.config_home());
     PluginManager::new(plugin_config)
-}
-
-fn resolve_plugin_path(cwd: &Path, config_home: &Path, value: &str) -> PathBuf {
-    let path = PathBuf::from(value);
-    if path.is_absolute() {
-        path
-    } else if value.starts_with('.') {
-        cwd.join(path)
-    } else {
-        config_home.join(path)
-    }
 }
 
 fn runtime_hook_config_from_plugin_hooks(
